requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
uvicorn==0.27.1
fastapi==0.109.2
pandas==2.2.0
openpyxl==3.1.2
pdf2docx==0.5.8
Pillow==10.2.0
pdfplumber==0.10.4
python-multipart==0.0.9
litellm
python-pptx==0.6.23
python-docx==1.1.0
curl_cffi
duckduckgo-search
PyJWT
passlib[bcrypt]
bcrypt
